@@ -1,17 +1,18 @@
 import setuptools
 
-with open('README.md', 'r') as f:
+with open("README.md", "r") as f:
     long_description = f.read()
 
 setuptools.setup(
-    name='ColBERT',
-    version='0.2.0',
-    author='Omar Khattab',
-    author_email='okhattab@stanford.edu',
+    name="ColBERT",
+    version="0.2.0",
+    author="Omar Khattab",
+    author_email="okhattab@stanford.edu",
     description="Efficient and Effective Passage Search via Contextualized Late Interaction over BERT",
     long_description=long_description,
-    long_description_content_type='text/markdown',
-    url='https://github.com/stanford-futuredata/ColBERT',
+    long_description_content_type="text/markdown",
+    url="https://github.com/stanford-futuredata/ColBERT",
     packages=setuptools.find_packages(),
-    python_requires='>=3.6',
+    include_package_data=True,
+    python_requires=">=3.8",
 )
