@@ -3,22 +3,23 @@ EVENTUALLY: Tune the batch sizes selected here for a good balance of speed and g
 """
 
 import os
-import torch
-import numpy as np
+import pathlib
 from itertools import product
 
-from colbert.infra.config import ColBERTConfig
+import numpy as np
+import torch
 from colbert.indexing.codecs.residual_embeddings import ResidualEmbeddings
+from colbert.infra.config import ColBERTConfig
 from colbert.utils.utils import print_message
-
-import pathlib
 from torch.utils.cpp_extension import load
 
 
 class ResidualCodec:
     Embeddings = ResidualEmbeddings
 
-    def __init__(self, config, centroids, avg_residual=None, bucket_cutoffs=None, bucket_weights=None):
+    def __init__(
+        self, config, centroids, avg_residual=None, bucket_cutoffs=None, bucket_weights=None
+    ):
         self.use_gpu = config.total_visible_gpus > 0
 
         ResidualCodec.try_load_torch_extensions(self.use_gpu)
@@ -44,7 +45,9 @@ class ResidualCodec:
         if not self.use_gpu and self.bucket_weights is not None:
             self.bucket_weights = self.bucket_weights.to(torch.float32)
 
-        self.arange_bits = torch.arange(0, self.nbits, device='cuda' if self.use_gpu else 'cpu', dtype=torch.uint8)
+        self.arange_bits = torch.arange(
+            0, self.nbits, device="cuda" if self.use_gpu else "cpu", dtype=torch.uint8
+        )
 
         self.rank = config.rank
 
@@ -63,7 +66,7 @@ class ResidualCodec:
                 # Reverse the endianness of each bit subsequence (e.g. 10 -> 01)
                 y = 0
                 for k in range(self.nbits - 1, -1, -1):
-                    y += ((x >> (self.nbits - k - 1)) & 1) * (2 ** k)
+                    y += ((x >> (self.nbits - k - 1)) & 1) * (2**k)
 
                 # Set the corresponding bits in the output byte
                 z |= y
@@ -76,17 +79,9 @@ class ResidualCodec:
         # given n bits per lookup
         keys_per_byte = 8 // self.nbits
         if self.bucket_weights is not None:
-            self.decompression_lookup_table = (
-                torch.tensor(
-                    list(
-                        product(
-                            list(range(len(self.bucket_weights))),
-                            repeat=keys_per_byte
-                        )
-                    )
-                )
-                .to(torch.uint8)
-            )
+            self.decompression_lookup_table = torch.tensor(
+                list(product(list(range(len(self.bucket_weights))), repeat=keys_per_byte))
+            ).to(torch.uint8)
         else:
             self.decompression_lookup_table = None
         if self.use_gpu:
@@ -99,31 +94,27 @@ class ResidualCodec:
         if hasattr(cls, "loaded_extensions") or not use_gpu:
             return
 
-        print_message(f"Loading decompress_residuals_cpp extension (set COLBERT_LOAD_TORCH_EXTENSION_VERBOSE=True for more info)...")
+        print_message(
+            f"Loading decompress_residuals_cpp extension (set COLBERT_LOAD_TORCH_EXTENSION_VERBOSE=True for more info)..."
+        )
         decompress_residuals_cpp = load(
             name="decompress_residuals_cpp",
             sources=[
-                os.path.join(
-                    pathlib.Path(__file__).parent.resolve(), "decompress_residuals.cpp"
-                ),
-                os.path.join(
-                    pathlib.Path(__file__).parent.resolve(), "decompress_residuals.cu"
-                ),
+                os.path.join(pathlib.Path(__file__).parent.resolve(), "decompress_residuals.cpp"),
+                os.path.join(pathlib.Path(__file__).parent.resolve(), "decompress_residuals.cu"),
             ],
             verbose=os.getenv("COLBERT_LOAD_TORCH_EXTENSION_VERBOSE", "False") == "True",
         )
         cls.decompress_residuals = decompress_residuals_cpp.decompress_residuals_cpp
 
-        print_message(f"Loading packbits_cpp extension (set COLBERT_LOAD_TORCH_EXTENSION_VERBOSE=True for more info)...")
+        print_message(
+            f"Loading packbits_cpp extension (set COLBERT_LOAD_TORCH_EXTENSION_VERBOSE=True for more info)..."
+        )
         packbits_cpp = load(
             name="packbits_cpp",
             sources=[
-                os.path.join(
-                    pathlib.Path(__file__).parent.resolve(), "packbits.cpp"
-                ),
-                os.path.join(
-                    pathlib.Path(__file__).parent.resolve(), "packbits.cu"
-                ),
+                os.path.join(pathlib.Path(__file__).parent.resolve(), "packbits.cpp"),
+                os.path.join(pathlib.Path(__file__).parent.resolve(), "packbits.cu"),
             ],
             verbose=os.getenv("COLBERT_LOAD_TORCH_EXTENSION_VERBOSE", "False") == "True",
         )
@@ -134,27 +125,33 @@ class ResidualCodec:
     @classmethod
     def load(cls, index_path):
         config = ColBERTConfig.load_from_index(index_path)
-        centroids_path = os.path.join(index_path, 'centroids.pt')
-        avgresidual_path = os.path.join(index_path, 'avg_residual.pt')
-        buckets_path = os.path.join(index_path, 'buckets.pt')
+        centroids_path = os.path.join(index_path, "centroids.pt")
+        avgresidual_path = os.path.join(index_path, "avg_residual.pt")
+        buckets_path = os.path.join(index_path, "buckets.pt")
 
-        centroids = torch.load(centroids_path, map_location='cpu')
-        avg_residual = torch.load(avgresidual_path, map_location='cpu')
-        bucket_cutoffs, bucket_weights = torch.load(buckets_path, map_location='cpu')
+        centroids = torch.load(centroids_path, map_location="cpu")
+        avg_residual = torch.load(avgresidual_path, map_location="cpu")
+        bucket_cutoffs, bucket_weights = torch.load(buckets_path, map_location="cpu")
 
         if avg_residual.dim() == 0:
             avg_residual = avg_residual.item()
 
-        return cls(config=config, centroids=centroids, avg_residual=avg_residual, bucket_cutoffs=bucket_cutoffs, bucket_weights=bucket_weights)
+        return cls(
+            config=config,
+            centroids=centroids,
+            avg_residual=avg_residual,
+            bucket_cutoffs=bucket_cutoffs,
+            bucket_weights=bucket_weights,
+        )
 
     def save(self, index_path):
         assert self.avg_residual is not None
         assert torch.is_tensor(self.bucket_cutoffs), self.bucket_cutoffs
         assert torch.is_tensor(self.bucket_weights), self.bucket_weights
 
-        centroids_path = os.path.join(index_path, 'centroids.pt')
-        avgresidual_path = os.path.join(index_path, 'avg_residual.pt')
-        buckets_path = os.path.join(index_path, 'buckets.pt')
+        centroids_path = os.path.join(index_path, "centroids.pt")
+        avgresidual_path = os.path.join(index_path, "avg_residual.pt")
+        buckets_path = os.path.join(index_path, "buckets.pt")
 
         torch.save(self.centroids.half(), centroids_path)
         torch.save((self.bucket_cutoffs, self.bucket_weights), buckets_path)
@@ -173,7 +170,7 @@ class ResidualCodec:
             codes_ = self.compress_into_codes(batch, out_device=batch.device)
             centroids_ = self.lookup_centroids(codes_, out_device=batch.device)
 
-            residuals_ = (batch - centroids_)
+            residuals_ = batch - centroids_
 
             codes.append(codes_.cpu())
             residuals.append(self.binarize(residuals_).cpu())
@@ -185,7 +182,9 @@ class ResidualCodec:
 
     def binarize(self, residuals):
         residuals = torch.bucketize(residuals.float(), self.bucket_cutoffs).to(dtype=torch.uint8)
-        residuals = residuals.unsqueeze(-1).expand(*residuals.size(), self.nbits)  # add a new nbits-wide dim
+        residuals = residuals.unsqueeze(-1).expand(
+            *residuals.size(), self.nbits
+        )  # add a new nbits-wide dim
         residuals = residuals >> self.arange_bits  # divide by 2^bit for each bit position
         residuals = residuals & 1  # apply mod 2 to binarize
 
@@ -203,8 +202,8 @@ class ResidualCodec:
 
     def compress_into_codes(self, embs, out_device):
         """
-            EVENTUALLY: Fusing the kernels or otherwise avoiding materalizing the entire matrix before max(dim=0)
-                        seems like it would help here a lot.
+        EVENTUALLY: Fusing the kernels or otherwise avoiding materalizing the entire matrix before max(dim=0)
+                    seems like it would help here a lot.
         """
 
         codes = []
@@ -212,18 +211,26 @@ class ResidualCodec:
         bsize = (1 << 29) // self.centroids.size(0)
         for batch in embs.split(bsize):
             if self.use_gpu:
-                indices = (self.centroids @ batch.T.cuda().half()).max(dim=0).indices.to(device=out_device)
+                indices = (
+                    (self.centroids @ batch.T.cuda().half())
+                    .max(dim=0)
+                    .indices.to(device=out_device)
+                )
             else:
-                indices = (self.centroids @ batch.T.cpu().float()).max(dim=0).indices.to(device=out_device)
+                indices = (
+                    (self.centroids @ batch.T.cpu().float())
+                    .max(dim=0)
+                    .indices.to(device=out_device)
+                )
             codes.append(indices)
 
         return torch.cat(codes)
 
     def lookup_centroids(self, codes, out_device):
         """
-            Handles multi-dimensional codes too.
+        Handles multi-dimensional codes too.
 
-            EVENTUALLY: The .split() below should happen on a flat view.
+        EVENTUALLY: The .split() below should happen on a flat view.
         """
 
         centroids = []
@@ -236,10 +243,10 @@ class ResidualCodec:
 
         return torch.cat(centroids)
 
-    #@profile
+    # @profile
     def decompress(self, compressed_embs: Embeddings):
         """
-            We batch below even if the target device is CUDA to avoid large temporary buffers causing OOM.
+        We batch below even if the target device is CUDA to avoid large temporary buffers causing OOM.
         """
 
         codes, residuals = compressed_embs.codes, compressed_embs.residuals
@@ -260,7 +267,7 @@ class ResidualCodec:
                 ).cuda()
             else:
                 # TODO: Remove dead code
-                centroids_ = self.lookup_centroids(codes_, out_device='cpu')
+                centroids_ = self.lookup_centroids(codes_, out_device="cpu")
                 residuals_ = self.reversed_bit_map[residuals_.long()]
                 residuals_ = self.decompression_lookup_table[residuals_.long()]
                 residuals_ = residuals_.reshape(residuals_.shape[0], -1)
