@@ -1,46 +1,46 @@
-from colbert.infra.config.config import ColBERTConfig
-from colbert.search.strided_tensor import StridedTensor
-from colbert.utils.utils import print_message, flatten
-from colbert.modeling.base_colbert import BaseColBERT
-
-import torch
-import string
-
 import os
 import pathlib
+import string
+
+import torch
+from colbert.infra.config.config import ColBERTConfig
+from colbert.modeling.base_colbert import BaseColBERT
+from colbert.search.strided_tensor import StridedTensor
+from colbert.utils.utils import flatten, print_message
 from torch.utils.cpp_extension import load
 
 
 class ColBERT(BaseColBERT):
     """
-        This class handles the basic encoding and scoring operations in ColBERT. It is used for training.
+    This class handles the basic encoding and scoring operations in ColBERT. It is used for training.
     """
 
-    def __init__(self, name='bert-base-uncased', colbert_config=None):
+    def __init__(self, name="bert-base-uncased", colbert_config=None):
         super().__init__(name, colbert_config)
         self.use_gpu = colbert_config.total_visible_gpus > 0
 
         ColBERT.try_load_torch_extensions(self.use_gpu)
 
         if self.colbert_config.mask_punctuation:
-            self.skiplist = {w: True
-                             for symbol in string.punctuation
-                             for w in [symbol, self.raw_tokenizer.encode(symbol, add_special_tokens=False)[0]]}
+            self.skiplist = {
+                w: True
+                for symbol in string.punctuation
+                for w in [symbol, self.raw_tokenizer.encode(symbol, add_special_tokens=False)[0]]
+            }
         self.pad_token = self.raw_tokenizer.pad_token_id
-
 
     @classmethod
     def try_load_torch_extensions(cls, use_gpu):
         if hasattr(cls, "loaded_extensions") or use_gpu:
             return
 
-        print_message(f"Loading segmented_maxsim_cpp extension (set COLBERT_LOAD_TORCH_EXTENSION_VERBOSE=True for more info)...")
+        print_message(
+            f"Loading segmented_maxsim_cpp extension (set COLBERT_LOAD_TORCH_EXTENSION_VERBOSE=True for more info)..."
+        )
         segmented_maxsim_cpp = load(
             name="segmented_maxsim_cpp",
             sources=[
-                os.path.join(
-                    pathlib.Path(__file__).parent.resolve(), "segmented_maxsim.cpp"
-                ),
+                os.path.join(pathlib.Path(__file__).parent.resolve(), "segmented_maxsim.cpp"),
             ],
             extra_cflags=["-O3"],
             verbose=os.getenv("COLBERT_LOAD_TORCH_EXTENSION_VERBOSE", "False") == "True",
@@ -51,7 +51,7 @@ class ColBERT(BaseColBERT):
 
     def forward(self, Q, D):
         Q = self.query(*Q)
-        D, D_mask = self.doc(*D, keep_dims='return_mask')
+        D, D_mask = self.doc(*D, keep_dims="return_mask")
 
         # Repeat each query encoding for every corresponding document.
         Q_duplicated = Q.repeat_interleave(self.colbert_config.nway, dim=0).contiguous()
@@ -65,14 +65,18 @@ class ColBERT(BaseColBERT):
 
     def compute_ib_loss(self, Q, D, D_mask):
         # TODO: Organize the code below! Quite messy.
-        scores = (D.unsqueeze(0) @ Q.permute(0, 2, 1).unsqueeze(1)).flatten(0, 1)  # query-major unsqueeze
+        scores = (D.unsqueeze(0) @ Q.permute(0, 2, 1).unsqueeze(1)).flatten(
+            0, 1
+        )  # query-major unsqueeze
 
         scores = colbert_score_reduce(scores, D_mask.repeat(Q.size(0), 1, 1), self.colbert_config)
 
         nway = self.colbert_config.nway
-        all_except_self_negatives = [list(range(qidx*D.size(0), qidx*D.size(0) + nway*qidx+1)) +
-                                     list(range(qidx*D.size(0) + nway * (qidx+1), qidx*D.size(0) + D.size(0)))
-                                     for qidx in range(Q.size(0))]
+        all_except_self_negatives = [
+            list(range(qidx * D.size(0), qidx * D.size(0) + nway * qidx + 1))
+            + list(range(qidx * D.size(0) + nway * (qidx + 1), qidx * D.size(0) + D.size(0)))
+            for qidx in range(Q.size(0))
+        ]
 
         scores = scores[flatten(all_except_self_negatives)]
         scores = scores.view(Q.size(0), -1)  # D.size(0) - self.colbert_config.nway + 1)
@@ -86,18 +90,25 @@ class ColBERT(BaseColBERT):
         Q = self.bert(input_ids, attention_mask=attention_mask)[0]
         Q = self.linear(Q)
 
-        mask = torch.tensor(self.mask(input_ids, skiplist=[]), device=self.device).unsqueeze(2).float()
+        mask = (
+            torch.tensor(self.mask(input_ids, skiplist=[]), device=self.device).unsqueeze(2).float()
+        )
         Q = Q * mask
 
         return torch.nn.functional.normalize(Q, p=2, dim=2)
 
     def doc(self, input_ids, attention_mask, keep_dims=True):
-        assert keep_dims in [True, False, 'return_mask']
+        assert keep_dims in [True, False, "return_mask"]
 
         input_ids, attention_mask = input_ids.to(self.device), attention_mask.to(self.device)
         D = self.bert(input_ids, attention_mask=attention_mask)[0]
         D = self.linear(D)
-        mask = torch.tensor(self.mask(input_ids, skiplist=self.skiplist), device=self.device).unsqueeze(2).float()
+
+        mask = (
+            torch.tensor(self.mask(input_ids, skiplist=self.skiplist), device=self.device)
+            .unsqueeze(2)
+            .float()
+        )
         D = D * mask
 
         D = torch.nn.functional.normalize(D, p=2, dim=2)
@@ -108,24 +119,34 @@ class ColBERT(BaseColBERT):
             D, mask = D.cpu(), mask.bool().cpu().squeeze(-1)
             D = [d[mask[idx]] for idx, d in enumerate(D)]
 
-        elif keep_dims == 'return_mask':
+        elif keep_dims == "return_mask":
             return D, mask.bool()
 
         return D
 
     def score(self, Q, D_padded, D_mask):
         # assert self.colbert_config.similarity == 'cosine'
-        if self.colbert_config.similarity == 'l2':
-            assert self.colbert_config.interaction == 'colbert'
-            return (-1.0 * ((Q.unsqueeze(2) - D_padded.unsqueeze(1))**2).sum(-1)).max(-1).values.sum(-1)
+
+        if self.colbert_config.similarity == "l2":
+            assert self.colbert_config.interaction == "colbert"
+            return (
+                (-1.0 * ((Q.unsqueeze(2) - D_padded.unsqueeze(1)) ** 2).sum(-1))
+                .max(-1)
+                .values.sum(-1)
+            )
+
         return colbert_score(Q, D_padded, D_mask, config=self.colbert_config)
 
     def mask(self, input_ids, skiplist):
-        mask = [[(x not in skiplist) and (x != self.pad_token) for x in d] for d in input_ids.cpu().tolist()]
+        mask = [
+            [(x not in skiplist) and (x != self.pad_token) for x in d]
+            for d in input_ids.cpu().tolist()
+        ]
         return mask
 
 
 # TODO: In Query/DocTokenizer, use colbert.raw_tokenizer
+
 
 # TODO: The masking below might also be applicable in the kNN part
 def colbert_score_reduce(scores_padded, D_mask, config: ColBERTConfig):
@@ -133,20 +154,20 @@ def colbert_score_reduce(scores_padded, D_mask, config: ColBERTConfig):
     scores_padded[D_padding] = -9999
     scores = scores_padded.max(1).values
 
-    assert config.interaction in ['colbert', 'flipr'], config.interaction
+    assert config.interaction in ["colbert", "flipr"], config.interaction
 
-    if config.interaction == 'flipr':
+    if config.interaction == "flipr":
         assert config.query_maxlen == 64, ("for now", config)
         # assert scores.size(1) == config.query_maxlen, scores.size()
 
         K1 = config.query_maxlen // 2
         K2 = 8
 
-        A = scores[:, :config.query_maxlen].topk(K1, dim=-1).values.sum(-1)
+        A = scores[:, : config.query_maxlen].topk(K1, dim=-1).values.sum(-1)
         B = 0
 
         if K2 <= scores.size(1) - config.query_maxlen:
-            B = scores[:, config.query_maxlen:].topk(K2, dim=-1).values.sum(1)
+            B = scores[:, config.query_maxlen :].topk(K2, dim=-1).values.sum(1)
 
         return A + B
 
@@ -156,11 +177,11 @@ def colbert_score_reduce(scores_padded, D_mask, config: ColBERTConfig):
 # TODO: Wherever this is called, pass `config=`
 def colbert_score(Q, D_padded, D_mask, config=ColBERTConfig()):
     """
-        Supply sizes Q = (1 | num_docs, *, dim) and D = (num_docs, *, dim).
-        If Q.size(0) is 1, the matrix will be compared with all passages.
-        Otherwise, each query matrix will be compared against the *aligned* passage.
+    Supply sizes Q = (1 | num_docs, *, dim) and D = (num_docs, *, dim).
+    If Q.size(0) is 1, the matrix will be compared with all passages.
+    Otherwise, each query matrix will be compared against the *aligned* passage.
 
-        EVENTUALLY: Consider masking with -inf for the maxsim (or enforcing a ReLU).
+    EVENTUALLY: Consider masking with -inf for the maxsim (or enforcing a ReLU).
     """
 
     use_gpu = config.total_visible_gpus > 0
@@ -178,7 +199,7 @@ def colbert_score(Q, D_padded, D_mask, config=ColBERTConfig()):
 
 def colbert_score_packed(Q, D_packed, D_lengths, config=ColBERTConfig()):
     """
-        Works with a single query only.
+    Works with a single query only.
     """
 
     use_gpu = config.total_visible_gpus > 0
@@ -194,7 +215,9 @@ def colbert_score_packed(Q, D_packed, D_lengths, config=ColBERTConfig()):
     scores = D_packed @ Q.to(dtype=D_packed.dtype).T
 
     if use_gpu or config.interaction == "flipr":
-        scores_padded, scores_mask = StridedTensor(scores, D_lengths, use_gpu=use_gpu).as_padded_tensor()
+        scores_padded, scores_mask = StridedTensor(
+            scores, D_lengths, use_gpu=use_gpu
+        ).as_padded_tensor()
 
         return colbert_score_reduce(scores_padded, scores_mask, config)
     else:
