@@ -1,6 +1,4 @@
-from .trainer import Trainer
 from .indexer import Indexer
-from .searcher import Searcher
-from .index_updater import IndexUpdater
-
 from .modeling.checkpoint import Checkpoint
+from .searcher import Searcher
+from .trainer import Trainer
