@@ -1,5 +1,4 @@
 from .collection import *
-from .queries import *
-
-from .ranking import *
 from .examples import *
+from .queries import *
+from .ranking import *
