@@ -1,10 +1,10 @@
-from colbert.infra.run import Run
 import os
-import ujson
 
-from colbert.utils.utils import print_message
+import ujson
 from colbert.infra.provenance import Provenance
-from utility.utils.save_metadata import get_metadata_only
+from colbert.infra.run import Run
+from colbert.utility.utils.save_metadata import get_metadata_only
+from colbert.utils.utils import print_message
 
 
 class Examples:
@@ -16,7 +16,7 @@ class Examples:
 
     def provenance(self):
         return self.__provenance
-    
+
     def toDict(self):
         return self.provenance()
 
@@ -41,27 +41,31 @@ class Examples:
 
         if rank or nranks:
             assert rank in range(nranks), (rank, nranks)
-            return [self.data[idx] for idx in range(0, len(self.data), nranks)]  # if line_idx % nranks == rank
+            return [
+                self.data[idx] for idx in range(0, len(self.data), nranks)
+            ]  # if line_idx % nranks == rank
 
         return list(self.data)
 
     def save(self, new_path):
-        assert 'json' in new_path.strip('/').split('/')[-1].split('.'), "TODO: Support .json[l] too."
+        assert "json" in new_path.strip("/").split("/")[-1].split(
+            "."
+        ), "TODO: Support .json[l] too."
 
         print_message(f"#> Writing {len(self.data) / 1000_000.0}M examples to {new_path}")
 
-        with Run().open(new_path, 'w') as f:
+        with Run().open(new_path, "w") as f:
             for example in self.data:
                 ujson.dump(example, f)
-                f.write('\n')
+                f.write("\n")
 
             output_path = f.name
             print_message(f"#> Saved examples with {len(self.data)} lines to {f.name}")
-        
-        with Run().open(f'{new_path}.meta', 'w') as f:
+
+        with Run().open(f"{new_path}.meta", "w") as f:
             d = {}
-            d['metadata'] = get_metadata_only()
-            d['provenance'] = self.provenance()
+            d["metadata"] = get_metadata_only()
+            d["provenance"] = self.provenance()
             line = ujson.dumps(d, indent=4)
             f.write(line)
 
