@@ -1,15 +1,15 @@
 import os
+
 import tqdm
 import ujson
 from colbert.infra.provenance import Provenance
-
 from colbert.infra.run import Run
-from colbert.utils.utils import print_message, groupby_first_item
-from utility.utils.save_metadata import get_metadata_only
+from colbert.utility.utils.save_metadata import get_metadata_only
+from colbert.utils.utils import groupby_first_item, print_message
 
 
 def numericize(v):
-    if '.' in v:
+    if "." in v:
         return float(v)
 
     return int(v)
@@ -19,7 +19,7 @@ def load_ranking(path):  # works with annotated and un-annotated ranked lists
     print_message("#> Loading the ranked lists from", path)
 
     with open(path) as f:
-        return [list(map(numericize, line.strip().split('\t'))) for line in f]
+        return [list(map(numericize, line.strip().split("\t"))) for line in f]
 
 
 class Ranking:
@@ -29,14 +29,16 @@ class Ranking:
 
     def provenance(self):
         return self.__provenance
-    
+
     def toDict(self):
-        return {'provenance': self.provenance()}
+        return {"provenance": self.provenance()}
 
     def _prepare_data(self, data):
         # TODO: Handle list of lists???
         if isinstance(data, dict):
-            self.flat_ranking = [(qid, *rest) for qid, subranking in data.items() for rest in subranking]
+            self.flat_ranking = [
+                (qid, *rest) for qid, subranking in data.items() for rest in subranking
+            ]
             return data
 
         self.flat_ranking = data
@@ -61,23 +63,25 @@ class Ranking:
         raise NotImplementedError
 
     def save(self, new_path):
-        assert 'tsv' in new_path.strip('/').split('/')[-1].split('.'), "TODO: Support .json[l] too."
+        assert "tsv" in new_path.strip("/").split("/")[-1].split("."), "TODO: Support .json[l] too."
 
-        with Run().open(new_path, 'w') as f:
+        with Run().open(new_path, "w") as f:
             for items in self.flat_ranking:
-                line = '\t'.join(map(lambda x: str(int(x) if type(x) is bool else x), items)) + '\n'
+                line = "\t".join(map(lambda x: str(int(x) if type(x) is bool else x), items)) + "\n"
                 f.write(line)
 
             output_path = f.name
-            print_message(f"#> Saved ranking of {len(self.data)} queries and {len(self.flat_ranking)} lines to {f.name}")
-        
-        with Run().open(f'{new_path}.meta', 'w') as f:
+            print_message(
+                f"#> Saved ranking of {len(self.data)} queries and {len(self.flat_ranking)} lines to {f.name}"
+            )
+
+        with Run().open(f"{new_path}.meta", "w") as f:
             d = {}
-            d['metadata'] = get_metadata_only()
-            d['provenance'] = self.provenance()
+            d["metadata"] = get_metadata_only()
+            d["provenance"] = self.provenance()
             line = ujson.dumps(d, indent=4)
             f.write(line)
-        
+
         return output_path
 
     @classmethod
