@@ -1,22 +1,23 @@
 import os
-import ujson
-import torch
 import random
+from collections import OrderedDict, defaultdict
 
-from collections import defaultdict, OrderedDict
-
-from colbert.parameters import DEVICE
+import torch
+import ujson
 from colbert.modeling.colbert import ColBERT
-from colbert.utils.utils import print_message, load_checkpoint
+from colbert.parameters import DEVICE
+from colbert.utils.utils import load_checkpoint, print_message
 
 
 def load_model(args, do_print=True):
-    colbert = ColBERT.from_pretrained('bert-base-uncased',
-                                      query_maxlen=args.query_maxlen,
-                                      doc_maxlen=args.doc_maxlen,
-                                      dim=args.dim,
-                                      similarity_metric=args.similarity,
-                                      mask_punctuation=args.mask_punctuation)
+    colbert = ColBERT.from_pretrained(
+        "bert-base-uncased",
+        query_maxlen=args.query_maxlen,
+        doc_maxlen=args.doc_maxlen,
+        dim=args.dim,
+        similarity_metric=args.similarity,
+        mask_punctuation=args.mask_punctuation,
+    )
     colbert = colbert.to(DEVICE)
 
     print_message("#> Loading model checkpoint.", condition=do_print)
