@@ -1,10 +1,9 @@
 import torch
-
 from colbert.infra.run import Run
-from colbert.utils.utils import print_message, batch
+from colbert.utils.utils import batch, print_message
 
 
-class CollectionEncoder():
+class CollectionEncoder:
     def __init__(self, config, checkpoint):
         self.config = config
         self.checkpoint = checkpoint
@@ -23,8 +22,12 @@ class CollectionEncoder():
             # Storing on the GPU helps with speed of masking, etc.
             # But ideally this batching happens internally inside docFromText.
             for passages_batch in batch(passages, self.config.bsize * 50):
-                embs_, doclens_ = self.checkpoint.docFromText(passages_batch, bsize=self.config.bsize,
-                                                              keep_dims='flatten', showprogress=(not self.use_gpu))
+                embs_, doclens_ = self.checkpoint.docFromText(
+                    passages_batch,
+                    bsize=self.config.bsize,
+                    keep_dims="flatten",
+                    showprogress=(not self.use_gpu),
+                )
                 embs.append(embs_)
                 doclens.extend(doclens_)
 
