@@ -10,6 +10,15 @@ class RunConfig(BaseConfig, RunSettings):
 
 
 @dataclass
-class ColBERTConfig(RunSettings, ResourceSettings, DocSettings, QuerySettings, TrainingSettings,
-                    IndexingSettings, SearchSettings, BaseConfig, TokenizerSettings):
+class ColBERTConfig(
+    RunSettings,
+    ResourceSettings,
+    DocSettings,
+    QuerySettings,
+    TrainingSettings,
+    IndexingSettings,
+    SearchSettings,
+    BaseConfig,
+    TokenizerSettings,
+):
     pass
