@@ -1,25 +1,19 @@
-import os
-import torch
-import ujson
 import dataclasses
-
-from typing import Any
+import os
 from collections import defaultdict
 from dataclasses import dataclass, fields
-from colbert.utils.utils import timestamp, torch_load_dnn
+from typing import Any
 
-from utility.utils.save_metadata import get_metadata_only
+import torch
+import ujson
+from colbert.utility.utils.save_metadata import get_metadata_only
+from colbert.utils.utils import timestamp, torch_load_dnn
 
 
 @dataclass
 class DefaultVal:
     val: Any
-    
-    def __hash__(self):
-        return hash(repr(self.val))
 
-    def __eq__(self, other):
-        self.val == other.val
 
 @dataclass
 class CoreConfig:
@@ -38,26 +32,25 @@ class CoreConfig:
 
             if not isinstance(field_val, DefaultVal):
                 self.assigned[field.name] = True
-    
+
     def assign_defaults(self):
         for field in fields(self):
             setattr(self, field.name, field.default.val)
             self.assigned[field.name] = True
 
     def configure(self, ignore_unrecognized=True, **kw_args):
-        ignored = set()
-
-        for key, value in kw_args.items():
-            self.set(key, value, ignore_unrecognized) or ignored.update({key})
-
-        return ignored
-
         """
         # TODO: Take a config object, not kw_args.
 
         for key in config.assigned:
             value = getattr(config, key)
         """
+        ignored = set()
+
+        for key, value in kw_args.items():
+            self.set(key, value, ignore_unrecognized) or ignored.update({key})
+
+        return ignored
 
     def set(self, key, value, ignore_unrecognized=False):
         if hasattr(self, key):
@@ -72,7 +65,7 @@ class CoreConfig:
         print(ujson.dumps(dataclasses.asdict(self), indent=4))
 
     def __export_value(self, v):
-        v = v.provenance() if hasattr(v, 'provenance') else v
+        v = v.provenance() if hasattr(v, "provenance") else v
 
         if isinstance(v, list) and len(v) > 100:
             v = (f"list with {len(v)} elements starting with...", v[:3])
