@@ -1,6 +1,6 @@
+import inspect
 import sys
 import traceback
-import inspect
 
 
 class Provenance:
@@ -13,7 +13,7 @@ class Provenance:
 
         for frame in trace[2:-1]:
             try:
-                frame = f'{frame.filename}:{frame.lineno}:{frame.function}:   {frame.code_context[0].strip()}'
+                frame = f"{frame.filename}:{frame.lineno}:{frame.function}:   {frame.code_context[0].strip()}"
                 output.append(frame)
             except:
                 output.append(None)
@@ -25,19 +25,19 @@ class Provenance:
         return dict(self.__dict__)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     p = Provenance()
     print(p.toDict().keys())
 
     import ujson
-    print(ujson.dumps(p, indent=4))
 
+    print(ujson.dumps(p, indent=4))
 
     class X:
         def __init__(self) -> None:
             pass
-        
+
         def toDict(self):
-            return {'key': 1}
-    
+            return {"key": 1}
+
     print(ujson.dumps(X()))
