@@ -1,11 +1,10 @@
 import random
 
-from colbert.utils.utils import print_message
-from utility.utils.save_metadata import save_metadata
-from utility.supervision.triples import sample_for_query
-
-from colbert.data.ranking import Ranking
 from colbert.data.examples import Examples
+from colbert.data.ranking import Ranking
+from colbert.utility.supervision.triples import sample_for_query
+from colbert.utility.utils.save_metadata import save_metadata
+from colbert.utils.utils import print_message
 
 MAX_NUM_TRIPLES = 40_000_000
 
@@ -24,12 +23,14 @@ class Triples:
 
         for processing_idx, qid in enumerate(self.qid2rankings):
             l = sample_for_query(qid, self.qid2rankings[qid], positives, depth, False, None)
-            NonEmptyQIDs += (len(l) > 0)
+            NonEmptyQIDs += len(l) > 0
             Triples.extend(l)
 
             if processing_idx % (10_000) == 0:
-                print_message(f"#> Done with {processing_idx+1} questions!\t\t "
-                              f"{str(len(Triples) / 1000)}k triples for {NonEmptyQIDs} unqiue QIDs.")
+                print_message(
+                    f"#> Done with {processing_idx+1} questions!\t\t "
+                    f"{str(len(Triples) / 1000)}k triples for {NonEmptyQIDs} unqiue QIDs."
+                )
 
         print_message(f"#> Sub-sample the triples (if > {MAX_NUM_TRIPLES})..")
         print_message(f"#> len(Triples) = {len(Triples)}")
@@ -49,4 +50,3 @@ class Triples:
         Examples(data=self.Triples).save(new_path)
 
         # save_metadata(f'{output}.meta', args)  # TODO: What args to save?? {seed, positives, depth, rankings if path or else whatever provenance the rankings object shares}
-
