@@ -1,12 +1,10 @@
 import os
+
 import torch
-import sys
-
-from colbert.utils.utils import torch_load_dnn
-
-from transformers import AutoTokenizer
-from colbert.modeling.hf_colbert import class_factory
 from colbert.infra.config import ColBERTConfig
+from colbert.modeling.hf_colbert import HF_ColBERT
+from colbert.utils.utils import torch_load_dnn
+from transformers import AutoTokenizer
 
 
 class BaseColBERT(torch.nn.Module):
@@ -17,23 +15,15 @@ class BaseColBERT(torch.nn.Module):
     Like HF, evaluation mode is the default.
     """
 
-    def __init__(self, name_or_path, colbert_config=None):
+    def __init__(self, name, colbert_config=None):
         super().__init__()
 
-        self.colbert_config = ColBERTConfig.from_existing(ColBERTConfig.load_from_checkpoint(name_or_path), colbert_config)
-        self.name = self.colbert_config.model_name or name_or_path
-
-        try:
-            HF_ColBERT = class_factory(self.name)
-        except:
-            self.name = 'bert-base-uncased' # TODO: Double check that this is appropriate here in all cases
-            HF_ColBERT = class_factory(self.name)
-
-        # assert self.name is not None
-        # HF_ColBERT = class_factory(self.name)
-        
-        self.model = HF_ColBERT.from_pretrained(name_or_path, colbert_config=self.colbert_config)
-        self.raw_tokenizer = AutoTokenizer.from_pretrained(name_or_path)
+        self.name = name
+        self.colbert_config = ColBERTConfig.from_existing(
+            ColBERTConfig.load_from_checkpoint(name), colbert_config
+        )
+        self.model = HF_ColBERT.from_pretrained(name, colbert_config=self.colbert_config)
+        self.raw_tokenizer = AutoTokenizer.from_pretrained(self.model.base)
 
         self.eval()
 
@@ -43,7 +33,7 @@ class BaseColBERT(torch.nn.Module):
 
     @property
     def bert(self):
-        return self.model.LM
+        return self.model.bert
 
     @property
     def linear(self):
@@ -54,7 +44,9 @@ class BaseColBERT(torch.nn.Module):
         return self.model.score_scaler
 
     def save(self, path):
-        assert not path.endswith('.dnn'), f"{path}: We reserve *.dnn names for the deprecated checkpoint format."
+        assert not path.endswith(
+            ".dnn"
+        ), f"{path}: We reserve *.dnn names for the deprecated checkpoint format."
 
         self.model.save_pretrained(path)
         self.raw_tokenizer.save_pretrained(path)
@@ -62,40 +54,43 @@ class BaseColBERT(torch.nn.Module):
         self.colbert_config.save_for_checkpoint(path)
 
 
-if __name__ == '__main__':
+if __name__ == "__main__":
     import random
-    import numpy as np
 
-    from colbert.infra.run import Run
+    import numpy as np
     from colbert.infra.config import RunConfig
+    from colbert.infra.run import Run
 
     random.seed(12345)
     np.random.seed(12345)
     torch.manual_seed(12345)
 
     with Run().context(RunConfig(gpus=2)):
-        m = BaseColBERT('bert-base-uncased', colbert_config=ColBERTConfig(Run().config, doc_maxlen=300, similarity='l2'))
+        m = BaseColBERT(
+            "bert-base-uncased",
+            colbert_config=ColBERTConfig(Run().config, doc_maxlen=300, similarity="l2"),
+        )
         m.colbert_config.help()
         print(m.linear.weight)
-        m.save('/future/u/okhattab/tmp/2021/08/model.deleteme2/')
+        m.save("/future/u/okhattab/tmp/2021/08/model.deleteme2/")
 
-    m2 = BaseColBERT('/future/u/okhattab/tmp/2021/08/model.deleteme2/')
+    m2 = BaseColBERT("/future/u/okhattab/tmp/2021/08/model.deleteme2/")
     m2.colbert_config.help()
     print(m2.linear.weight)
 
     exit()
 
-    m = BaseColBERT('/future/u/okhattab/tmp/2021/08/model.deleteme/')
-    print('BaseColBERT', m.linear.weight)
-    print('BaseColBERT', m.colbert_config)
+    m = BaseColBERT("/future/u/okhattab/tmp/2021/08/model.deleteme/")
+    print("BaseColBERT", m.linear.weight)
+    print("BaseColBERT", m.colbert_config)
 
     exit()
 
     # m = HF_ColBERT.from_pretrained('nreimers/MiniLMv2-L6-H768-distilled-from-BERT-Large')
-    m = HF_ColBERT.from_pretrained('/future/u/okhattab/tmp/2021/08/model.deleteme/')
-    print('HF_ColBERT', m.linear.weight)
+    m = HF_ColBERT.from_pretrained("/future/u/okhattab/tmp/2021/08/model.deleteme/")
+    print("HF_ColBERT", m.linear.weight)
 
-    m.save_pretrained('/future/u/okhattab/tmp/2021/08/model.deleteme/')
+    m.save_pretrained("/future/u/okhattab/tmp/2021/08/model.deleteme/")
 
     # old = OldColBERT.from_pretrained('bert-base-uncased')
     # print(old.bert.encoder.layer[10].attention.self.value.weight)
@@ -105,14 +100,15 @@ if __name__ == '__main__':
     # torch.manual_seed(12345)
 
     dnn = torch_load_dnn(
-        "/future/u/okhattab/root/TACL21/experiments/Feb26.NQ/train.py/ColBERT.C3/checkpoints/colbert-60000.dnn")
+        "/future/u/okhattab/root/TACL21/experiments/Feb26.NQ/train.py/ColBERT.C3/checkpoints/colbert-60000.dnn"
+    )
     # base = dnn.get('arguments', {}).get('model', 'bert-base-uncased')
 
     # new = BaseColBERT.from_pretrained('bert-base-uncased', state_dict=dnn['model_state_dict'])
 
     # print(new.bert.encoder.layer[10].attention.self.value.weight)
 
-    print(dnn['model_state_dict']['linear.weight'])
+    print(dnn["model_state_dict"]["linear.weight"])
     # print(dnn['model_state_dict']['bert.encoder.layer.10.attention.self.value.weight'])
 
     # # base_model_prefix
