@@ -1,13 +1,13 @@
 import os
-import ujson
-import torch
-import numpy as np
-import tqdm
 
-from colbert.utils.utils import lengths2offsets, print_message, dotdict, flatten
+import numpy as np
+import torch
+import tqdm
+import ujson
 from colbert.indexing.codecs.residual import ResidualCodec
 from colbert.indexing.utils import optimize_ivf
 from colbert.search.strided_tensor import StridedTensor
+from colbert.utils.utils import dotdict, flatten, lengths2offsets, print_message
 
 
 class IndexLoader:
@@ -29,10 +29,14 @@ class IndexLoader:
         print_message(f"#> Loading IVF...")
 
         if os.path.exists(os.path.join(self.index_path, "ivf.pid.pt")):
-            ivf, ivf_lengths = torch.load(os.path.join(self.index_path, "ivf.pid.pt"), map_location='cpu')
+            ivf, ivf_lengths = torch.load(
+                os.path.join(self.index_path, "ivf.pid.pt"), map_location="cpu"
+            )
         else:
             assert os.path.exists(os.path.join(self.index_path, "ivf.pt"))
-            ivf, ivf_lengths = torch.load(os.path.join(self.index_path, "ivf.pt"), map_location='cpu')
+            ivf, ivf_lengths = torch.load(
+                os.path.join(self.index_path, "ivf.pt"), map_location="cpu"
+            )
             ivf, ivf_lengths = optimize_ivf(ivf, ivf_lengths, self.index_path)
 
         if False:
@@ -50,22 +54,23 @@ class IndexLoader:
         print_message("#> Loading doclens...")
 
         for chunk_idx in tqdm.tqdm(range(self.num_chunks)):
-            with open(os.path.join(self.index_path, f'doclens.{chunk_idx}.json')) as f:
+            with open(os.path.join(self.index_path, f"doclens.{chunk_idx}.json")) as f:
                 chunk_doclens = ujson.load(f)
                 doclens.extend(chunk_doclens)
 
         self.doclens = torch.tensor(doclens)
 
     def _load_embeddings(self):
-        self.embeddings = ResidualCodec.Embeddings.load_chunks(self.index_path, range(self.num_chunks),
-                                                               self.num_embeddings)
+        self.embeddings = ResidualCodec.Embeddings.load_chunks(
+            self.index_path, range(self.num_chunks), self.num_embeddings
+        )
 
     @property
     def metadata(self):
         try:
             self._metadata
         except:
-            with open(os.path.join(self.index_path, 'metadata.json')) as f:
+            with open(os.path.join(self.index_path, "metadata.json")) as f:
                 self._metadata = ujson.load(f)
 
         return self._metadata
@@ -77,10 +82,9 @@ class IndexLoader:
     @property
     def num_chunks(self):
         # EVENTUALLY: If num_chunks doesn't exist (i.e., old index), fall back to counting doclens.*.json files.
-        return self.metadata['num_chunks']
+        return self.metadata["num_chunks"]
 
     @property
     def num_embeddings(self):
         # EVENTUALLY: If num_embeddings doesn't exist (i.e., old index), sum the values in doclens.*.json files.
-        return self.metadata['num_embeddings']
-
+        return self.metadata["num_embeddings"]
