@@ -1,27 +1,28 @@
 import os
-import ujson
-
 from functools import partial
-from colbert.infra.config.config import ColBERTConfig
-from colbert.utils.utils import print_message, zipstar
-from colbert.modeling.tokenization import QueryTokenizer, DocTokenizer, tensorize_triples
-from colbert.evaluation.loaders import load_collection
 
+import ujson
 from colbert.data.collection import Collection
-from colbert.data.queries import Queries
 from colbert.data.examples import Examples
+from colbert.data.queries import Queries
+from colbert.evaluation.loaders import load_collection
+from colbert.infra.config.config import ColBERTConfig
+from colbert.modeling.tokenization import DocTokenizer, QueryTokenizer, tensorize_triples
+from colbert.utils.utils import print_message, zipstar
 
 # from colbert.utils.runs import Run
 
 
-class LazyBatcher():
+class LazyBatcher:
     def __init__(self, config: ColBERTConfig, triples, queries, collection, rank=0, nranks=1):
         self.bsize, self.accumsteps = config.bsize, config.accumsteps
         self.nway = config.nway
 
         self.query_tokenizer = QueryTokenizer(config)
         self.doc_tokenizer = DocTokenizer(config)
-        self.tensorize_triples = partial(tensorize_triples, self.query_tokenizer, self.doc_tokenizer)
+        self.tensorize_triples = partial(
+            tensorize_triples, self.query_tokenizer, self.doc_tokenizer
+        )
         self.position = 0
 
         self.triples = Examples.cast(triples, nway=self.nway).tolist(rank, nranks)
@@ -39,13 +40,14 @@ class LazyBatcher():
         self.position = endpos
 
         if offset + self.bsize > len(self.triples):
-            raise StopIteration
+            self.position = 0
+            return self.__next__()
 
         all_queries, all_passages, all_scores = [], [], []
 
         for position in range(offset, endpos):
             query, *pids = self.triples[position]
-            pids = pids[:self.nway]
+            pids = pids[: self.nway]
 
             query = self.queries[query]
 
@@ -59,7 +61,7 @@ class LazyBatcher():
             all_queries.append(query)
             all_passages.extend(passages)
             all_scores.extend(scores)
-        
+
         assert len(all_scores) in [0, len(all_passages)], len(all_scores)
 
         return self.collate(all_queries, all_passages, all_scores)
@@ -68,7 +70,9 @@ class LazyBatcher():
         assert len(queries) == self.bsize
         assert len(passages) == self.nway * self.bsize
 
-        return self.tensorize_triples(queries, passages, scores, self.bsize // self.accumsteps, self.nway)
+        return self.tensorize_triples(
+            queries, passages, scores, self.bsize // self.accumsteps, self.nway
+        )
 
     # def skip_to_batch(self, batch_idx, intended_batch_size):
     #     Run.warn(f'Skipping to batch #{batch_idx} (with intended_batch_size = {intended_batch_size}) for training.')
