@@ -1,7 +1,8 @@
 import os
 import random
-import torch
+
 import numpy as np
+import torch
 
 ALREADY_INITALIZED = False
 
@@ -9,9 +10,9 @@ ALREADY_INITALIZED = False
 
 
 def init(rank):
-    nranks = 'WORLD_SIZE' in os.environ and int(os.environ['WORLD_SIZE'])
+    nranks = "WORLD_SIZE" in os.environ and int(os.environ["WORLD_SIZE"])
     nranks = max(1, nranks)
-    is_distributed = (nranks > 1) or ('WORLD_SIZE' in os.environ)
+    is_distributed = (nranks > 1) or ("WORLD_SIZE" in os.environ)
 
     global ALREADY_INITALIZED
     if ALREADY_INITALIZED:
@@ -21,16 +22,16 @@ def init(rank):
 
     if is_distributed and torch.cuda.is_available():
         num_gpus = torch.cuda.device_count()
-        print(f'nranks = {nranks} \t num_gpus = {num_gpus} \t device={rank % num_gpus}')
+        print(f"nranks = {nranks} \t num_gpus = {num_gpus} \t device={rank % num_gpus}")
 
         torch.cuda.set_device(rank % num_gpus)
-        torch.distributed.init_process_group(backend='nccl', init_method='env://')
+        torch.distributed.init_process_group(backend="nccl", init_method="env://")
 
     return nranks, is_distributed
 
 
 def barrier(rank):
-    nranks = 'WORLD_SIZE' in os.environ and int(os.environ['WORLD_SIZE'])
+    nranks = "WORLD_SIZE" in os.environ and int(os.environ["WORLD_SIZE"])
     nranks = max(1, nranks)
 
     if rank >= 0 and nranks > 1:
