@@ -1,21 +1,20 @@
-import os
-import tqdm
-import torch
 import datetime
 import itertools
-
-from multiprocessing import Pool
+import os
 from collections import OrderedDict, defaultdict
+from multiprocessing import Pool
+
+import torch
+import tqdm
 
 
 def print_message(*s, condition=True, pad=False):
-    s = ' '.join([str(x) for x in s])
+    s = " ".join([str(x) for x in s])
     msg = "[{}] {}".format(datetime.datetime.now().strftime("%b %d, %H:%M:%S"), s)
 
     if condition:
-        msg = msg if not pad else f'\n{msg}\n'
+        msg = msg if not pad else f"\n{msg}\n"
         print(msg, flush=True)
-
 
     return msg
 
@@ -39,24 +38,25 @@ def file_tqdm(file):
 
 def torch_load_dnn(path):
     if path.startswith("http:") or path.startswith("https:"):
-        dnn = torch.hub.load_state_dict_from_url(path, map_location='cpu')
+        dnn = torch.hub.load_state_dict_from_url(path, map_location="cpu")
     else:
-        dnn = torch.load(path, map_location='cpu')
-    
+        dnn = torch.load(path, map_location="cpu")
+
     return dnn
+
 
 def save_checkpoint(path, epoch_idx, mb_idx, model, optimizer, arguments=None):
     print(f"#> Saving a checkpoint to {path} ..")
 
-    if hasattr(model, 'module'):
+    if hasattr(model, "module"):
         model = model.module  # extract model from a distributed/data-parallel wrapper
 
     checkpoint = {}
-    checkpoint['epoch'] = epoch_idx
-    checkpoint['batch'] = mb_idx
-    checkpoint['model_state_dict'] = model.state_dict()
-    checkpoint['optimizer_state_dict'] = optimizer.state_dict()
-    checkpoint['arguments'] = arguments
+    checkpoint["epoch"] = epoch_idx
+    checkpoint["batch"] = mb_idx
+    checkpoint["model_state_dict"] = model.state_dict()
+    checkpoint["optimizer_state_dict"] = optimizer.state_dict()
+    checkpoint["arguments"] = arguments
 
     torch.save(checkpoint, path)
 
@@ -69,48 +69,49 @@ def load_checkpoint(path, model, checkpoint=None, optimizer=None, do_print=True)
         checkpoint = load_checkpoint_raw(path)
 
     try:
-        model.load_state_dict(checkpoint['model_state_dict'])
+        model.load_state_dict(checkpoint["model_state_dict"])
     except:
         print_message("[WARNING] Loading checkpoint with strict=False")
-        model.load_state_dict(checkpoint['model_state_dict'], strict=False)
+        model.load_state_dict(checkpoint["model_state_dict"], strict=False)
 
     if optimizer:
-        optimizer.load_state_dict(checkpoint['optimizer_state_dict'])
+        optimizer.load_state_dict(checkpoint["optimizer_state_dict"])
 
     if do_print:
-        print_message("#> checkpoint['epoch'] =", checkpoint['epoch'])
-        print_message("#> checkpoint['batch'] =", checkpoint['batch'])
+        print_message("#> checkpoint['epoch'] =", checkpoint["epoch"])
+        print_message("#> checkpoint['batch'] =", checkpoint["batch"])
 
     return checkpoint
 
 
 def load_checkpoint_raw(path):
     if path.startswith("http:") or path.startswith("https:"):
-        checkpoint = torch.hub.load_state_dict_from_url(path, map_location='cpu')
+        checkpoint = torch.hub.load_state_dict_from_url(path, map_location="cpu")
     else:
-        checkpoint = torch.load(path, map_location='cpu')
+        checkpoint = torch.load(path, map_location="cpu")
 
-    state_dict = checkpoint['model_state_dict']
+    state_dict = checkpoint["model_state_dict"]
     new_state_dict = OrderedDict()
     for k, v in state_dict.items():
         name = k
-        if k[:7] == 'module.':
+        if k[:7] == "module.":
             name = k[7:]
         new_state_dict[name] = v
 
-    checkpoint['model_state_dict'] = new_state_dict
+    checkpoint["model_state_dict"] = new_state_dict
 
     return checkpoint
 
 
 def create_directory(path):
     if os.path.exists(path):
-        print('\n')
-        print_message("#> Note: Output directory", path, 'already exists\n\n')
+        print("\n")
+        print_message("#> Note: Output directory", path, "already exists\n\n")
     else:
-        print('\n')
-        print_message("#> Creating directory", path, '\n\n')
+        print("\n")
+        print_message("#> Creating directory", path, "\n\n")
         os.makedirs(path)
+
 
 # def batch(file, bsize):
 #     while True:
@@ -131,7 +132,7 @@ def f7(seq):
 def batch(group, bsize, provide_offset=False):
     offset = 0
     while offset < len(group):
-        L = group[offset: offset + bsize]
+        L = group[offset : offset + bsize]
         yield ((offset, L) if provide_offset else L)
         offset += len(L)
     return
@@ -142,6 +143,7 @@ class dotdict(dict):
     dot.notation access to dictionary attributes
     Credit: derek73 @ https://stackoverflow.com/questions/2352181
     """
+
     __getattr__ = dict.__getitem__
     __setattr__ = dict.__setitem__
     __delattr__ = dict.__delitem__
@@ -193,10 +195,11 @@ def zip_first(L1, L2):
 
 
 def int_or_float(val):
-    if '.' in val:
+    if "." in val:
         return float(val)
-        
+
     return int(val)
+
 
 def load_ranking(path, types=None, lazy=False):
     print_message(f"#> Loading the ranked lists from {path} ..")
@@ -209,8 +212,10 @@ def load_ranking(path, types=None, lazy=False):
             types = itertools.cycle([int_or_float])
 
         with open(path) as f:
-            lists = [[typ(x) for typ, x in zip_first(types, line.strip().split('\t'))]
-                     for line in file_tqdm(f)]
+            lists = [
+                [typ(x) for typ, x in zip_first(types, line.strip().split("\t"))]
+                for line in file_tqdm(f)
+            ]
 
     return lists
 
@@ -236,7 +241,7 @@ def groupby_first_item(lst):
 
 def process_grouped_by_first_item(lst):
     """
-        Requires items in list to already be grouped by first item.
+    Requires items in list to already be grouped by first item.
     """
 
     groups = defaultdict(list)
@@ -284,8 +289,10 @@ def lengths2offsets(lengths):
 class NullContextManager(object):
     def __init__(self, dummy_resource=None):
         self.dummy_resource = dummy_resource
+
     def __enter__(self):
         return self.dummy_resource
+
     def __exit__(self, *args):
         pass
 
@@ -302,9 +309,9 @@ def load_batch_backgrounds(args, qids):
         if len(back) and type(back[0]) == int:
             x = [args.collection[pid] for pid in back]
         else:
-            x = [args.collectionX.get(pid, '') for pid in back]
+            x = [args.collectionX.get(pid, "") for pid in back]
 
-        x = ' [SEP] '.join(x)
+        x = " [SEP] ".join(x)
         qbackgrounds.append(x)
-    
+
     return qbackgrounds
